- fixed compilation ( std::max() used on non-identical types) - fixed harmful warning (taking address of a temporary)
git-svn-id: https://bioproj.extra.cea.fr/neurosvn/brainvisa/cortical_surface/cortical_surface-private/trunk@39747 613412ce-1361-4b7a-93bf-a3c75af1bae4

+++ src/library/geodesicpath/geodesic_algorithm_exact.h
@@ -549,7 +549,8 @@ inline void GeodesicAlgorithmExact::propagate(std::vector<SurfacePoint>& sources
 
 	while(!m_queue.empty())
 	{
-		m_queue_max_size = std::max(m_queue.size(), m_queue_max_size);
+		m_queue_max_size = std::max(m_queue.size(),
+                                            (size_t) m_queue_max_size);
 
 		unsigned const check_period = 10;
     	if(++m_iterations % check_period == 0)		//check if we covered all required vertices
